update for new didethresolver format
