+++ gateway-types/src/lib.rs
@@ -1,6 +1,3 @@
-use lib_didethresolver::types::{
-    Attribute, KeyEncoding, KeyMetadata, KeyPurpose, KeyType, PublicKey,
-};
 use serde::{Deserialize, Serialize};
 
 /// Address of the did:ethr Registry on Sepolia
@@ -20,23 +17,4 @@ pub struct Message {
     pub r: Vec<u8>,
     /// Signature of S
     pub s: Vec<u8>,
-}
-
-/// The XMTP-specific attribute type
-#[derive(Serialize, Deserialize)]
-pub enum XmtpAttributeType {
-    InstallationKey,
-}
-
-impl From<XmtpAttributeType> for Attribute {
-    fn from(attribute: XmtpAttributeType) -> Self {
-        match attribute {
-            XmtpAttributeType::InstallationKey => Attribute::PublicKey(PublicKey {
-                key_type: KeyType::Ed25519VerificationKey2020,
-                purpose: KeyPurpose::Xmtp,
-                encoding: KeyEncoding::Hex,
-                metadata: Some(KeyMetadata::Installation),
-            }),
-        }
-    }
 }
